# [CARBONDATA-2018][DataLoad] Optimization in reading/writing for sort temp row