Pick up the no-sort fields in the row and pack them as bytes array and skip parsing them during merge sort to reduce CPU consumption

I've tested it in my cluster and seen about 8% performance gained (74MB/s/Node -> 81MB/s/Node) in data loading. Please note that global_sort will not gain benefit from this feature since there are no sort temp file in that procedure.

Be sure to do all of the following checklist to help us incorporate 
your contribution quickly and easily:

 - [x] Any interfaces changed?
 `Some internal used interface has been changed`
 - [x] Any backward compatibility impacted?
 `No`
 - [x] Document update required?
`No`
 - [x] Testing done
        Please provide details on 
        - Whether new unit test cases have been added or why no new tests are required?
`No`
        - How it is tested? Please attach test report.
`Tested in 3-node cluster with real business data`
        - Is it a performance related change? Please attach the performance test report.
`Yes, I've tested it in my cluster and seen about 8% performance gained (74MB/s/Node -> 81MB/s/Node) in data loading.`
        - Any additional information to help reviewers in testing this change.
`No`
 - [x] For large changes, please consider breaking it into sub-tasks under an umbrella JIRA. 
`Unrelated`