@xuchuanyin 2018-02-08
this PR depends on #1952